requirements.watcher.txt
python-frontmatter
watchdog
python-dotenv
pymongo
PyYAML
permit
langchain-openai